+++ alerts_in_ua/__init__.py
@@ -1,7 +1,5 @@
-__version__ = "0.2.7"
+__version__ = "1.0.0"
 
-from .async_client import AsyncClient
-from .client import Client
-from .location_uid_resolver import LocationUidResolver
+from alerts_in_ua.client import Client
 
-__all__ = ['Client', 'AsyncClient']
+__all__ = ('Client',)
