+++ alerts_in_ua/client.py
@@ -1,33 +1,45 @@
-from typing import Union
+import os
 
+import aiohttp
 import requests
+from aiohttp import ClientTimeout
+from typing_extensions import Optional, Any, Dict, Union
 
-from .air_raid_alert_oblast_status import AirRaidAlertOblastStatus
-from .air_raid_alert_oblast_statuses import AirRaidAlertOblastStatuses
-from .alerts import Alerts
-from .errors import UnauthorizedError, RateLimitError, InternalServerError, ForbiddenError, ApiError
-from .location_uid_resolver import LocationUidResolver
-from .user_agent import UserAgent
+from alerts_in_ua.alerts import Alerts, HistoryPeriod, AirRaidOblastStatus, AirRaidOblastStatuses
+from alerts_in_ua.errors import (ApiError, RateLimitError,
+                                 ForbiddenError, UnauthorizedError, InternalServerError)
+from alerts_in_ua.location_resolver import resolve_uid, resolve_title
+from alerts_in_ua.user_agent import USER_AGENT
+
+REQUEST_TIMEOUT = 5
+API_BASE_URL = "https://api.alerts.in.ua"
 
 
 class Client:
-    REQUEST_TIMEOUT = 5
-    API_BASE_URL = "https://api.alerts.in.ua"
-
-    def __init__(self, token: str):
-        self.token = token
-        self.base_url = Client.API_BASE_URL + "/v1/"
-        self.location_uid_resolver = LocationUidResolver()
+    def __init__(self, token: Optional[str] = None):
+        self.token = self._check_token(token)
+        self.base_url = API_BASE_URL + "/v1/"
         self.headers = {
             "Accept": "application/json",
             "Authorization": f"Bearer {self.token}",
-            "User-Agent": UserAgent.get_user_agent()
+            "User-Agent": USER_AGENT
         }
-        self.cache = {}
+        self.cache: Dict[str, Any] = {}
 
-    def _request(self, endpoint: str, use_cache=True):
+    @staticmethod
+    def _check_token(token: Optional[str] = None) -> str:
+        if not token:
+            if 'AIU_API_TOKEN' in os.environ:
+                return os.environ['AIU_API_TOKEN']
+            raise EnvironmentError(
+                "API token not provided. Please put `token` to Client constructor "
+                "or set it as environment variable `AIU_API_TOKEN`"
+            )
+        return token
+
+    def _request(self, endpoint: str, use_cache: bool = True) -> Any:
         # Check if endpoint is in cache and return cached data if not modified
-        if endpoint in self.cache and use_cache == True:
+        if use_cache and self.cache.get(endpoint):
             cached_data = self.cache[endpoint]
             headers = {
                 **self.headers,
@@ -36,7 +48,7 @@ class Client:
             response = requests.get(
                 self.base_url + endpoint,
                 headers=headers,
-                timeout=Client.REQUEST_TIMEOUT,
+                timeout=REQUEST_TIMEOUT,
             )
             if response.status_code == 304:
                 return cached_data["Data"]
@@ -45,74 +57,170 @@ class Client:
         response = requests.get(
             self.base_url + endpoint,
             headers=self.headers,
-            timeout=Client.REQUEST_TIMEOUT,
+            timeout=REQUEST_TIMEOUT,
         )
 
         # Check if response is successful
-        if response.status_code == 200:
+        status_code = response.status_code
+        if status_code == 200:
             data = response.json()
             self.cache[endpoint] = {
                 "Data": data,
                 "Last-Modified": response.headers["Last-Modified"],
             }
             return data
-        else:
+        message = None
+        try:
+            data = response.json()  # can raise requests.exceptions.JSONDecodeError
+            json_message = data.get("message")
+            message = f"{json_message} HTTP Code: {status_code}"
+        finally:
+            self._process_exception(status_code, message)
+
+    async def _async_request(self, endpoint: str, use_cache: bool = True) -> Any:
+        async with aiohttp.ClientSession() as session:
+            # Check if endpoint is in cache and return cached data if not modified
+            if use_cache and self.cache.get(endpoint):
+                cached_data = self.cache[endpoint]
+                headers = {
+                    **self.headers,
+                    **{"If-Modified-Since": cached_data["Last-Modified"]},
+                }
+                response = await session.get(
+                    self.base_url + endpoint,
+                    headers=headers,
+                    timeout=ClientTimeout(REQUEST_TIMEOUT),
+                )
+                if response.status == 304:
+                    return cached_data["Data"]
+
+            # Make the request
+            response = await session.get(
+                self.base_url + endpoint,
+                headers=self.headers,
+                timeout=ClientTimeout(REQUEST_TIMEOUT),
+            )
+
+            # Check if response is successful
+            status_code = response.status
+            if response.status == 200:
+                data = await response.json()
+                self.cache[endpoint] = {
+                    "Data": data,
+                    "Last-Modified": response.headers["Last-Modified"],
+                }
+                return data
             message = None
             try:
-                data = response.json()
+                data = await response.json()  # can raise requests.exceptions.JSONDecodeError
                 json_message = data.get("message")
-                message = f"{json_message} HTTP Code:{response.status_code}"
-            except:
-                pass
-            if response.status_code == 401:
-                if message is None:
-                    message = "Unauthorized: Incorrect token"
-                raise UnauthorizedError(message)
-            elif response.status_code == 403:
-                if message is None:
-                    message = "Forbidden. API may not be available in some regions. Please ask api@alerts.in.ua for details."
-                raise ForbiddenError(message)
-            elif response.status_code == 429:
-                if message is None:
-                    message = "Too many requests: Rate limit exceeded"
-                raise RateLimitError(message)
-            elif response.status_code == 500:
-                raise InternalServerError("Internal server error")
-            else:
-                raise ApiError(f"Unknown error. HTTP Code:{response.status_code}")
+                message = f"{json_message} HTTP Code: {status_code}"
+            finally:
+                self._process_exception(status_code, message)
 
-    def get_active_alerts(self, use_cache=True) -> Alerts:
+    @staticmethod
+    def _process_exception(status_code, message):
+        if status_code == 401:
+            raise UnauthorizedError(message)
+        elif status_code == 403:
+            raise ForbiddenError(message)
+        elif status_code == 429:
+            raise RateLimitError(message)
+        elif status_code == 500:
+            raise InternalServerError(message)
+        else:
+            raise ApiError(f"Unknown error. HTTP Code:{status_code}")
+
+    def get_active(self, use_cache=True) -> Alerts:
         data = self._request("alerts/active.json", use_cache=use_cache)
-        return Alerts(data)
+        return Alerts(**data)
 
-    def get_alerts_history(self, oblast_uid_or_location_title: Union[int, str], period: str = 'week_ago',
-                           use_cache: bool = True) -> Alerts:
+    async def async_get_active(self, use_cache=True) -> Alerts:
+        data = await self._async_request("alerts/active.json", use_cache=use_cache)
+        return Alerts(**data)
+
+    def get_history(self, oblast_uid_or_location_title: Union[int, str],
+                    period: Union[HistoryPeriod, str] = HistoryPeriod.WEEK_AGO,
+                    use_cache: bool = True) -> Alerts:
         if isinstance(oblast_uid_or_location_title, str):
             if oblast_uid_or_location_title.isdigit():
                 oblast_uid = int(oblast_uid_or_location_title)
             else:
-                oblast_uid = self.location_uid_resolver.resolve_uid(oblast_uid_or_location_title)
-        else:
+                oblast_uid = resolve_uid(oblast_uid_or_location_title)
+        elif isinstance(oblast_uid_or_location_title, int):
             oblast_uid = oblast_uid_or_location_title
+        else:
+            raise TypeError(f"Expected an int or str, got {type(oblast_uid_or_location_title)}")
 
         url = f"regions/{oblast_uid}/alerts/{period}.json"
         data = self._request(url, use_cache=use_cache)
-        return Alerts(data)
+        return Alerts(**data)
 
-    def get_air_raid_alert_status(self, oblast_uid_or_location_title: Union[int, str], oblast_level_only=False,
-                                  use_cache=True) -> AirRaidAlertOblastStatus:
+    async def async_get_history(self, oblast_uid_or_location_title: Union[int, str],
+                                period: Union[HistoryPeriod, str] = HistoryPeriod.WEEK_AGO,
+                                use_cache: bool = True) -> Alerts:
         if isinstance(oblast_uid_or_location_title, str):
             if oblast_uid_or_location_title.isdigit():
                 oblast_uid = int(oblast_uid_or_location_title)
             else:
-                oblast_uid = self.location_uid_resolver.resolve_uid(oblast_uid_or_location_title)
+                oblast_uid = resolve_uid(oblast_uid_or_location_title)
         else:
             oblast_uid = oblast_uid_or_location_title
-        data = self._request(f"iot/active_air_raid_alerts/{oblast_uid}.json", use_cache=use_cache)
-        return AirRaidAlertOblastStatus(location_title=self.location_uid_resolver.resolve_location_title(oblast_uid),
-                                        status=data, oblast_level_only=oblast_level_only)
+        url = f"regions/{oblast_uid}/alerts/{period}.json"
+        data = await self._async_request(url, use_cache=use_cache)
+        return Alerts(**data)
 
-    def get_air_raid_alert_statuses_by_oblast(self, oblast_level_only=False,
-                                              use_cache=True) -> AirRaidAlertOblastStatuses:
-        data = self._request("iot/active_air_raid_alerts_by_oblast.json", use_cache=use_cache)
-        return AirRaidAlertOblastStatuses(data, oblast_level_only=oblast_level_only)
+    def get_air_raid(self, oblast_uid_or_location_title: Union[int, str],
+                     use_cache=True) -> AirRaidOblastStatus:
+        if isinstance(oblast_uid_or_location_title, str):
+            if oblast_uid_or_location_title.isdigit():
+                oblast_uid = int(oblast_uid_or_location_title)
+            else:
+                oblast_uid = resolve_uid(oblast_uid_or_location_title)
+        else:
+            oblast_uid = oblast_uid_or_location_title
+        data = self._request(
+            f"iot/active_air_raid_alerts/{oblast_uid}.json",
+            use_cache=use_cache
+        )
+        return AirRaidOblastStatus(
+            location_title=resolve_title(oblast_uid),
+            location_uid=oblast_uid,
+            status=data
+        )
+
+    async def async_get_air_raid(self, oblast_uid_or_location_title: Union[int, str],
+                                 use_cache=True) -> AirRaidOblastStatus:
+        if isinstance(oblast_uid_or_location_title, str):
+            if oblast_uid_or_location_title.isdigit():
+                oblast_uid = int(oblast_uid_or_location_title)
+            else:
+                oblast_uid = resolve_uid(oblast_uid_or_location_title)
+        else:
+            oblast_uid = oblast_uid_or_location_title
+        data = await self._async_request(
+            f"iot/active_air_raid_alerts/{oblast_uid}.json",
+            use_cache=use_cache
+        )
+        return AirRaidOblastStatus(
+            location_title=resolve_title(oblast_uid),
+            location_uid=oblast_uid,
+            status=data
+        )
+
+    def get_air_raids(self, use_cache=True) -> AirRaidOblastStatuses:
+        data = self._request(
+            "iot/active_air_raid_alerts_by_oblast.json",
+            use_cache=use_cache
+        )
+        return AirRaidOblastStatuses(data)
+
+    async def async_get_air_raids(self, use_cache=True) -> AirRaidOblastStatuses:
+        data = await self._async_request(
+            "iot/active_air_raid_alerts_by_oblast.json",
+            use_cache=use_cache
+        )
+        return AirRaidOblastStatuses(data)
+
+
+__all__ = ('Client',)
