--- alerts_in_ua_/__init__.py
@@ -1,5 +1,0 @@
-__version__ = "0.2.7"
-
-from alerts_in_ua_.client import Client
-
-__all__ = ('Client', )
