--- alerts_in_ua_/user_agent.py
@@ -1,6 +1,0 @@
-from alerts_in_ua_ import __version__
-
-# default const
-USER_AGENT = f"aiu-py-client/{__version__} (+https://alerts.in.ua)"
-
-__all__ = ("USER_AGENT",)
