+++ setup.py
@@ -1,28 +1,3 @@
-from setuptools import setup, find_packages
-from alerts_in_ua import __version__
-import pathlib
-import re
+from setuptools import setup
 
-here = pathlib.Path(__file__).parent.resolve()  # current path
-long_description = (here / 'README.md').read_text(encoding='utf-8')  # Get the long description from the README file
-
-setup(
-    name='alerts_details-in-ua',
-    version=__version__,
-    author='Ukrzen Team',
-    author_email='api@alerts_details.in.ua',
-    url='https://github.com/alerts-ua/alerts-in-ua-py',
-    description='Python library for alerts_details.in.ua API',
-    long_description=long_description,
-    long_description_content_type="text/markdown",
-    packages=find_packages(),
-    install_requires=[
-        'aiohttp', 'requests','pytz'
-    ],
-    classifiers=[
-        'License :: OSI Approved :: MIT License',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
-        'Programming Language :: Python :: 3.8',
-    ],
-)
+setup()
